backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.7
langgraph==0.2.38
langchain-community==0.3.7
langchain-groq==0.2.3
langchain-ollama==0.2.1
langchain-chroma==0.1.4
chromadb==0.5.15
sentence-transformers==3.2.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
typing-extensions==4.12.2
aiofiles==24.1.0
